wifi: enable trx bound and move idle core to CPU6
To improve the long DPC time on 4390, we need to define Tx/Rx tasklet bound to limit the executing packet counts for each DPC.
Since the this might caused wifi performance drop, we need to switch small affinity to CPU6 to improve performance as well.
And with CL pa/2797073, the long ISR issue while using CPU6 will also be fixed.

Bug: 283029172
Bug: 333488872
Test: Astress/SST/Performance/Regression test
Signed-off-by: Hana Huang <hanahuang@google.com>
Change-Id: Ie7a10f8abfe4b0f9ab37d3ad7453772fce20ca73