bcmdhd: Fix no roam scan case in AP+STA concurrency
Issue:
Framework HAL based roam disable/enable back fail to enable back roam when AP+STA is operational

Rootcause:
When the HAL API gets called for roam enable, there was a concurrency check which was causing the roam to not get enabled and also to not cache the framework roam enable state.

Fix:
Fixed the condition check in the HAL roam enable API path for the AP+STA concurrency case so that roam gets enabled.

Bug:338700123
Test:
1. enable sta+sap on phone
2. halutil_brcm -set_fw_roaming_state 0
3. halutil_brcm -set_fw_roaming_state 1
4. do the roaming test

Change-Id: Idca921ad57fcdc9eace47b0b22ce00e7a64dc4d8
Signed-off-by: Jithu Jance <jithu.jance@broadcom.corp-partner.google.com>